+++ lib/app.dart
@@ -1,9 +1,8 @@
-import 'dart:html' as html;
-import 'package:firebase/firebase.dart' as firebase;
 import 'package:dashboard/model.dart' as model;
-import 'package:dashboard/firebase.dart' as fb;
-import 'package:dashboard/utils.dart' as utils;
+import 'package:dashboard/path.dart' as path;
 import 'package:dashboard/navbar/controller.dart' as nav_controller;
+import 'package:dashboard/login/controller.dart' as login_controller;
+import 'package:dashboard/view.dart' as app_view;
 import 'package:dashboard/logger.dart';
 
 Logger logger = Logger('app.dart');
@@ -13,25 +12,32 @@ var links = [
   model.Link('/settings', 'Settings'),
 ];
 
-html.ButtonElement get loginButton => html.querySelector('#login-button');
-
 class App {
+  app_view.View view;
+  login_controller.Controller loginController;
   nav_controller.Controller navController;
 
   App() {
-    initFirebase();
+    loginController =
+        login_controller.Controller(_onSignInComplete, _onSignOutComplete);
     // TO THINK ABOUT: navController might want to take in the respective controllers as well
     navController = nav_controller.Controller(links);
-    utils.listenToPathChanges(_handlePath);
+    path.listenToChanges(_handlePath);
   }
 
-  void _loadData() async {
+  void _onSignInComplete() async {
+    view.showLoadingIndicator();
     logger.debug('Loading all data ...');
+    view.hideLoadingIndicator();
+  }
+
+  void _onSignOutComplete() {
+    logger.debug('Clearing all data ...');
   }
 
   void _handlePath(_) {
-    logger.debug('updated url to ${utils.currentPathname}');
-    switch (utils.currentPathname) {
+    logger.debug('updated url to ${path.currentName}');
+    switch (path.currentName) {
       case '/charts':
         logger.debug('Render charts');
         break;
@@ -39,49 +45,8 @@ class App {
         logger.debug('Render settings');
         break;
       default:
-        logger.error(
-            'Route ${utils.currentPathname} not handled, showing 404 page');
+        logger.error('Route ${path.currentName} not handled, showing 404 page');
         break;
     }
-  }
-
-  void initFirebase() async {
-    await fb.init('firebase/constants.json');
-    utils.setLoginDomains(fb.allowedEmailDomains);
-    fb.firebaseAuth.onAuthStateChanged.listen(_fbAuthChanged);
-    loginButton.onClick.listen((_) {
-      utils.disableLoginButton();
-      fb.signInWithGoogle();
-    });
-  }
-
-  void _fbAuthChanged(firebase.User user) async {
-    utils.enableLoginButton();
-    if (user == null) {
-      logger.debug('User not signed in');
-      utils.showLoginModal();
-      return;
-    }
-
-    if (!fb.allowedEmailDomains.any((domain) => user.email.endsWith(domain))) {
-      logger.error('Email domain not allowed');
-      await fb.deleteUser();
-      utils.showLoginError('Email domain not allowed');
-      return;
-    }
-
-    if (!user.emailVerified) {
-      logger.error('Email not verified');
-      await fb.deleteUser();
-      utils.showLoginError('Email is not verified');
-      return;
-    }
-
-    logger.debug('Loggedin as ${user.email}');
-    utils.hideLoginError();
-    utils.hideLoginModal();
-
-    await _loadData();
-    _handlePath(null);
   }
 }
