+++ lib/navbar/controller.dart
@@ -1,7 +1,7 @@
 import 'package:dashboard/logger.dart';
 import 'package:dashboard/navbar/view.dart' as nav_view;
 import 'package:dashboard/model.dart' as model;
-import 'package:dashboard/utils.dart' as utils;
+import 'package:dashboard/path.dart' as path;
 
 Logger logger = Logger('navbar/controller.dart');
 
@@ -26,7 +26,7 @@ class Controller {
   void _initialRender() {
     for (var link in _links) {
       _view.appendNavLink(
-          link.pathname, link.label, link.pathname == utils.currentPathname);
+          link.pathname, link.label, link.pathname == path.currentName);
     }
   }
 
@@ -34,7 +34,7 @@ class Controller {
     switch (action) {
       case UIAction.gotoURL:
         URLData url = data;
-        utils.gotoPath(url.pathname);
+        path.goto(url.pathname);
         _view.setNavlinkSelected(url.pathname);
         break;
       default:
