--- web/firebase/constants-example.json
@@ -1,11 +1,0 @@
-{
-  "apiKey": "fb_config",
-  "authDomain": "fb_config",
-  "databaseURL": "fb_config",
-  "projectId": "fb_config",
-  "storageBucket": "fb_config",
-  "messagingSenderId": "fb_config",
-  "appId": "fb_config",
-  "measurementId": "fb_config",
-  "allowedEmailDomains": "domain1.com|domain2.com"
-}
